Simulations Cont
This version adds the H_subj estimation of the simulated data.
The simulation filenames are different in this version and do not include underscores.
(1) This version of fitAdaptivityModel.m is made specifically for the simulated data. This version of getDataInfo.m has been tailored for the simulations as well.
(2)  Added the plotSimResult file so that it plots the results of the H_subj estimation.
(3) The file generateSimData.m facilitates the generation of all the simulation files as well as create the subj_info file in Data/Simulated_Analysis.