Fix bugs, change simulation filenames
There were two main bugs in the simulation workflow:
(1) In createStruct_sim.m, corrected the calculation of the log likelihood ratio.
(2) In fitAdaptivityModel_err.m, the calculation of choicehat was corrected so that it uses Lprior rather than Lposterior. The definition of e was simplified, but not changed.
(3) In generateSimData, simulations using hSubj of 0 or 1 were removed

The simulation file name changes were minor and affected the following files:
(1) createStruct_Sim.m
(2) generateSimData.m
(3) plotSimResult.R

Finally, added plotting of percent correct in the file plotSimResult.R.